requirements.txt
certifi==2022.6.15
configparser==5.2.0
influxdb-client==1.30.0
loguru==0.6.0
pip==22.0.4
PySnooper==1.1.1
python-dateutil==2.8.2
pyzmq==23.2.0
Rx==3.2.0
setuptools==58.1.0
six==1.16.0
urllib3==1.26.10
